api/requirements.txt
Flask
flask-marshmallow
flask-restful-swagger-3
python-dotenv
gunicorn
Flask-PyMongo
pymongo[srv]
pytest